progress on making SDK
