+++ hyperdbg/include/Definition.h
@@ -45,125 +45,6 @@ typedef RFLAGS * PRFLAGS;
 #define PDBEX_DEFAULT_CONFIGURATION "-j- -k- -e n -i"
 
 //////////////////////////////////////////////////
-//				Message Tracing                 //
-//////////////////////////////////////////////////
-
-/**
- * @brief Default buffer count of packets for message tracing
- * @details number of packets storage for regualr buffers
- */
-#define MaximumPacketsCapacity 1000
-
-/**
- * @brief Default buffer count of packets for message tracing
- * @details number of packets storage for priority buffers
- */
-#define MaximumPacketsCapacityPriority 10
-
-/**
- * @brief Size of each packet
- * @details NOTE : REMEMBER TO CHANGE IT IN USER-MODE APP TOO
- * @warning we redefine it on ScriptEngineEval.h change it on
- * that file too
- */
-#define PacketChunkSize 4096 // PAGE_SIZE
-
-/**
- * @brief size of user-mode buffer
- * @details Because of Opeation code at the start of the
- * buffer + 1 for null-termminating
- *
- */
-#define UsermodeBufferSize sizeof(UINT32) + PacketChunkSize + 1
-
-/**
- * @brief size of buffer for serial
- * @details the maximum packet size for sending over serial
- * User-mode buffer size + Header Structure Size + Count Of End Buffer Bytes
- *
- */
-#define MaxSerialPacketSize                               \
-    UsermodeBufferSize + sizeof(DEBUGGER_REMOTE_PACKET) + \
-        SERIAL_END_OF_BUFFER_CHARS_COUNT
-
-/**
- * @brief Final storage size of message tracing
- *
- */
-#define LogBufferSize \
-    MaximumPacketsCapacity *(PacketChunkSize + sizeof(BUFFER_HEADER))
-
-/**
- * @brief Final storage size of message tracing
- *
- */
-#define LogBufferSizePriority \
-    MaximumPacketsCapacityPriority *(PacketChunkSize + sizeof(BUFFER_HEADER))
-
-/**
- * @brief limitation of Windows DbgPrint message size
- * @details currently is not functional
- *
- */
-#define DbgPrintLimitation 512
-
-/**
- * @brief The seeds that user-mode codes use as the starter
- * of their events' tag
- *
- */
-#define DebuggerEventTagStartSeed 0x1000000
-
-/**
- * @brief The seeds that user-mode thread detail token start with it
- * @details This seed should not start with zero (0), otherwise it's
- * interpreted as error
- */
-#define DebuggerThreadDebuggingTagStartSeed 0x1000000
-
-/**
- * @brief The seeds that user-mode codes use as the starter
- * of their output source tag
- *
- */
-#define DebuggerOutputSourceTagStartSeed 0x1
-
-/**
- * @brief Determines how many sources a debugger can have for
- * a single event
- *
- */
-#define DebuggerOutputSourceMaximumRemoteSourceForSingleEvent 0x5
-
-//////////////////////////////////////////////////
-//               Remote Connection              //
-//////////////////////////////////////////////////
-
-/**
- * @brief default port of HyperDbg for listening by
- * debuggee (server, guest)
- *
- */
-#define DEFAULT_PORT "50000"
-
-/**
- * @brief Packet size for TCP connections
- * @details Note that we might add something to the kernel buffers
- * that's why we add 0x100 to it
- */
-#define COMMUNICATION_BUFFER_SIZE PacketChunkSize + 0x100
-
-//////////////////////////////////////////////////
-//            Breakpoint Backup                 //
-//////////////////////////////////////////////////
-
-/**
- * @brief maximum number of buffers to be allocated for a single
- * breakpoint
- */
-#define MAXIMUM_BREAKPOINTS_WITHOUT_CONTINUE 50
-
-//////////////////////////////////////////////////
 //                Config File                  //
 //////////////////////////////////////////////////
 
@@ -201,45 +82,6 @@ typedef RFLAGS * PRFLAGS;
 #define MAXIMUM_CHARACTER_FOR_OS_NAME 256
 
 //////////////////////////////////////////////////
-//             Operation Codes                  //
-//////////////////////////////////////////////////
-
-/**
- * @brief If a operation use this bit in its Operation code,
- * then it means that the operation should be performed
- * mandatorily in debuggee and should not be sent to the debugger
- */
-#define OPERATION_MANDATORY_DEBUGGEE_BIT (1 << 31)
-
-/**
- * @brief Message logs id that comes from kernel-mode to
- * user-mode
- * @details Message area >= 0x5
- */
-#define OPERATION_LOG_INFO_MESSAGE          0x1
-#define OPERATION_LOG_WARNING_MESSAGE       0x2
-#define OPERATION_LOG_ERROR_MESSAGE         0x3
-#define OPERATION_LOG_NON_IMMEDIATE_MESSAGE 0x4
-#define OPERATION_LOG_WITH_TAG              0x5
-
-#define OPERATION_COMMAND_FROM_DEBUGGER_CLOSE_AND_UNLOAD_VMM \
-    0x6 | OPERATION_MANDATORY_DEBUGGEE_BIT
-#define OPERATION_DEBUGGEE_USER_INPUT     0x7 | OPERATION_MANDATORY_DEBUGGEE_BIT
-#define OPERATION_DEBUGGEE_REGISTER_EVENT 0x8 | OPERATION_MANDATORY_DEBUGGEE_BIT
-#define OPERATION_DEBUGGEE_ADD_ACTION_TO_EVENT \
-    0x9 | OPERATION_MANDATORY_DEBUGGEE_BIT
-#define OPERATION_DEBUGGEE_CLEAR_EVENTS 0xa | OPERATION_MANDATORY_DEBUGGEE_BIT
-#define OPERATION_HYPERVISOR_DRIVER_IS_SUCCESSFULLY_LOADED \
-    0xb | OPERATION_MANDATORY_DEBUGGEE_BIT
-#define OPERATION_HYPERVISOR_DRIVER_END_OF_IRPS \
-    0xc | OPERATION_MANDATORY_DEBUGGEE_BIT
-#define OPERATION_COMMAND_FROM_DEBUGGER_RELOAD_SYMBOL \
-    0xd | OPERATION_MANDATORY_DEBUGGEE_BIT
-
-#define OPERATION_NOTIFICATION_FROM_USER_DEBUGGER_PAUSE \
-    0xe | OPERATION_MANDATORY_DEBUGGEE_BIT
-
-//////////////////////////////////////////////////
 //				   Test Cases                   //
 //////////////////////////////////////////////////
 
@@ -267,116 +109,6 @@ typedef RFLAGS * PRFLAGS;
  * @brief Maximum buffer to communicate between debugger and debuggee process
  */
 #define TEST_CASE_MAXIMUM_BUFFERS_TO_COMMUNICATE sizeof(DEBUGGEE_KERNEL_AND_USER_TEST_INFORMATION) * TEST_CASE_MAXIMUM_NUMBER_OF_KERNEL_TEST_CASES
-
-//////////////////////////////////////////////////
-//		Debugger Synchronization Objects        //
-//////////////////////////////////////////////////
-
-/**
- * @brief maximum number of event handles in kernel-debugger
- */
-#define DEBUGGER_MAXIMUM_SYNCRONIZATION_KERNEL_DEBUGGER_OBJECTS 0x40
-
-/**
- * @brief An event to show whether the debugger is running
- * or not in kernel-debugger
- *
- */
-
-//
-// Kernel-debugger
-//
-#define DEBUGGER_SYNCRONIZATION_OBJECT_KERNEL_DEBUGGER_IS_DEBUGGER_RUNNING                 0x0
-#define DEBUGGER_SYNCRONIZATION_OBJECT_KERNEL_DEBUGGER_STARTED_PACKET_RECEIVED             0x1
-#define DEBUGGER_SYNCRONIZATION_OBJECT_KERNEL_DEBUGGER_PAUSED_DEBUGGEE_DETAILS             0x2
-#define DEBUGGER_SYNCRONIZATION_OBJECT_KERNEL_DEBUGGER_CORE_SWITCHING_RESULT               0x3
-#define DEBUGGER_SYNCRONIZATION_OBJECT_KERNEL_DEBUGGER_PROCESS_SWITCHING_RESULT            0x4
-#define DEBUGGER_SYNCRONIZATION_OBJECT_KERNEL_DEBUGGER_THREAD_SWITCHING_RESULT             0x5
-#define DEBUGGER_SYNCRONIZATION_OBJECT_KERNEL_DEBUGGER_SCRIPT_RUNNING_RESULT               0x6
-#define DEBUGGER_SYNCRONIZATION_OBJECT_KERNEL_DEBUGGER_SCRIPT_FORMATS_RESULT               0x7
-#define DEBUGGER_SYNCRONIZATION_OBJECT_KERNEL_DEBUGGER_DEBUGGEE_FINISHED_COMMAND_EXECUTION 0x8
-#define DEBUGGER_SYNCRONIZATION_OBJECT_KERNEL_DEBUGGER_FLUSH_RESULT                        0x9
-#define DEBUGGER_SYNCRONIZATION_OBJECT_KERNEL_DEBUGGER_REGISTER_EVENT                      0xa
-#define DEBUGGER_SYNCRONIZATION_OBJECT_KERNEL_DEBUGGER_ADD_ACTION_TO_EVENT                 0xb
-#define DEBUGGER_SYNCRONIZATION_OBJECT_KERNEL_DEBUGGER_MODIFY_AND_QUERY_EVENT              0xc
-#define DEBUGGER_SYNCRONIZATION_OBJECT_KERNEL_DEBUGGER_READ_REGISTERS                      0xd
-#define DEBUGGER_SYNCRONIZATION_OBJECT_KERNEL_DEBUGGER_BP                                  0xe
-#define DEBUGGER_SYNCRONIZATION_OBJECT_KERNEL_DEBUGGER_LIST_OR_MODIFY_BREAKPOINTS          0xf
-#define DEBUGGER_SYNCRONIZATION_OBJECT_KERNEL_DEBUGGER_READ_MEMORY                         0x10
-#define DEBUGGER_SYNCRONIZATION_OBJECT_KERNEL_DEBUGGER_EDIT_MEMORY                         0x11
-#define DEBUGGER_SYNCRONIZATION_OBJECT_KERNEL_DEBUGGER_SYMBOL_RELOAD                       0x12
-#define DEBUGGER_SYNCRONIZATION_OBJECT_KERNEL_DEBUGGER_TEST_QUERY                          0x13
-#define DEBUGGER_SYNCRONIZATION_OBJECT_KERNEL_DEBUGGER_CALLSTACK_RESULT                    0x14
-#define DEBUGGER_SYNCRONIZATION_OBJECT_KERNEL_DEBUGGER_SEARCH_QUERY_RESULT                 0x15
-#define DEBUGGER_SYNCRONIZATION_OBJECT_KERNEL_DEBUGGER_VA2PA_AND_PA2VA_RESULT              0x16
-#define DEBUGGER_SYNCRONIZATION_OBJECT_KERNEL_DEBUGGER_PTE_RESULT                          0x17
-
-/**
- * @brief Maximum number of event handles in user-debugger
- */
-#define DEBUGGER_MAXIMUM_SYNCRONIZATION_USER_DEBUGGER_OBJECTS 0x40
-
-/**
- * @brief An event to show whether the debugger is running
- * or not in user-debugger
- *
- */
-
-//
-// User-debugger
-//
-#define DEBUGGER_SYNCRONIZATION_OBJECT_USER_DEBUGGER_IS_DEBUGGER_RUNNING 0x30
-
-//////////////////////////////////////////////////
-//            End of Buffer Detection           //
-//////////////////////////////////////////////////
-
-/**
- * @brief count of characters for serial end of buffer
- */
-#define SERIAL_END_OF_BUFFER_CHARS_COUNT 0x4
-
-/**
- * @brief characters of the buffer that we set at the end of
- * buffers for serial
- */
-#define SERIAL_END_OF_BUFFER_CHAR_1 0x00
-#define SERIAL_END_OF_BUFFER_CHAR_2 0x80
-#define SERIAL_END_OF_BUFFER_CHAR_3 0xEE
-#define SERIAL_END_OF_BUFFER_CHAR_4 0xFF
-
-/**
- * @brief count of characters for tcp end of buffer
- */
-#define TCP_END_OF_BUFFER_CHARS_COUNT 0x4
-
-/**
- * @brief characters of the buffer that we set at the end of
- * buffers for tcp
- */
-#define TCP_END_OF_BUFFER_CHAR_1 0x10
-#define TCP_END_OF_BUFFER_CHAR_2 0x20
-#define TCP_END_OF_BUFFER_CHAR_3 0x33
-#define TCP_END_OF_BUFFER_CHAR_4 0x44
-
-//////////////////////////////////////////////////
-//              Debugger Internals              //
-//////////////////////////////////////////////////
-
-/**
- * @brief debug register for step-over
- */
-#define DEBUGGER_DEBUG_REGISTER_FOR_STEP_OVER 0
-
-/**
- * @brief debug register to monitor thread changes
- */
-#define DEBUGGER_DEBUG_REGISTER_FOR_THREAD_MANAGEMENT 1
-
-/**
- * @brief debug register get entrypoint of user-mode process
- */
-#define DEBUGGER_DEBUG_REGISTER_FOR_USER_MODE_ENTRY_POINT 1
 
 //////////////////////////////////////////////////
 //              Symbols Details                 //
@@ -436,20 +168,6 @@ typedef struct _USERMODE_LOADED_MODULE_DETAILS
 
 } USERMODE_LOADED_MODULE_DETAILS, *PUSERMODE_LOADED_MODULE_DETAILS;
 
-//////////////////////////////////////////////////
-//              Processor Details               //
-//////////////////////////////////////////////////
-
-/**
- * @brief maximum instruction size in intel
- */
-#define MAXIMUM_INSTR_SIZE 16
-
-/**
- * @brief maximum size for call instruction in Intel
- */
-#define MAXIMUM_CALL_INSTR_SIZE 7
-
 /**
  * @brief Callback type that should be used to add
  * list of Addresses to ObjectNames
@@ -458,159 +176,22 @@ typedef struct _USERMODE_LOADED_MODULE_DETAILS
 typedef VOID (*SymbolMapCallback)(UINT64 Address, char * ModuleName, char * ObjectName, unsigned int ObjectSize);
 
 //////////////////////////////////////////////////
-//               Event Details                  //
+//              Processor Details               //
 //////////////////////////////////////////////////
 
 /**
- * @brief Integer gp registers (this structure is defined in
- * two places, make sure to change it in two places)
- *
+ * @brief maximum instruction size in Intel
  */
-#ifndef GUEST_REGS_DEFINED
-#    define GUEST_REGS_DEFINED
-
-//
-// DO NOT FUCKING TOUCH THIS STRUCTURE WITHOUT COORDINATION WITH SINA
-//
-typedef struct GUEST_REGS
-{
-    //
-    // DO NOT FUCKING TOUCH THIS STRUCTURE WITHOUT COORDINATION WITH SINA
-    //
-
-    UINT64 rax; // 0x00
-    UINT64 rcx; // 0x08
-    UINT64 rdx; // 0x10
-    UINT64 rbx; // 0x18
-    UINT64 rsp; // 0x20
-    UINT64 rbp; // 0x28
-    UINT64 rsi; // 0x30
-    UINT64 rdi; // 0x38
-    UINT64 r8;  // 0x40
-    UINT64 r9;  // 0x48
-    UINT64 r10; // 0x50
-    UINT64 r11; // 0x58
-    UINT64 r12; // 0x60
-    UINT64 r13; // 0x68
-    UINT64 r14; // 0x70
-    UINT64 r15; // 0x78
-
-    //
-    // DO NOT FUCKING TOUCH THIS STRUCTURE WITHOUT COORDINATION WITH SINA
-    //
-
-} GUEST_REGS, *PGUEST_REGS;
-#endif
+#define MAXIMUM_INSTR_SIZE 16
 
 /**
- * @brief struct for extra registers
- *
+ * @brief maximum size for call instruction in Intel
  */
-typedef struct GUEST_EXTRA_REGISTERS
-{
-    UINT16 CS;
-    UINT16 DS;
-    UINT16 FS;
-    UINT16 GS;
-    UINT16 ES;
-    UINT16 SS;
-    UINT64 RFLAGS;
-    UINT64 RIP;
-} GUEST_EXTRA_REGISTERS, *PGUEST_EXTRA_REGISTERS;
+#define MAXIMUM_CALL_INSTR_SIZE 7
 
-/**
- * @brief RFLAGS in structure format
- *
- */
-// typedef union _RFLAGS
-//{
-//     struct
-//     {
-//         UINT64 CarryFlag : 1;
-//         UINT64 ReadAs1 : 1;
-//         UINT64 ParityFlag : 1;
-//         UINT64 Reserved1 : 1;
-//         UINT64 AuxiliaryCarryFlag : 1;
-//         UINT64 Reserved2 : 1;
-//         UINT64 ZeroFlag : 1;
-//         UINT64 SignFlag : 1;
-//         UINT64 TrapFlag : 1;
-//         UINT64 InterruptEnableFlag : 1;
-//         UINT64 DirectionFlag : 1;
-//         UINT64 OverflowFlag : 1;
-//         UINT64 IoPrivilegeLevel : 2;
-//         UINT64 NestedTaskFlag : 1;
-//         UINT64 Reserved3 : 1;
-//         UINT64 ResumeFlag : 1;
-//         UINT64 Virtual8086ModeFlag : 1;
-//         UINT64 AlignmentCheckFlag : 1;
-//         UINT64 VirtualInterruptFlag : 1;
-//         UINT64 VirtualInterruptPendingFlag : 1;
-//         UINT64 IdentificationFlag : 1;
-//     } Fields;
-//
-//     UINT64 Value;
-// } RFLAGS, *PRFLAGS;
-
-/**
- * @brief enum to show type of all HyperDbg events
- *
- */
-typedef enum _DEBUGGER_EVENT_TYPE_ENUM
-{
-
-    HIDDEN_HOOK_READ_AND_WRITE,
-    HIDDEN_HOOK_READ,
-    HIDDEN_HOOK_WRITE,
-
-    HIDDEN_HOOK_EXEC_DETOURS,
-    HIDDEN_HOOK_EXEC_CC,
-
-    SYSCALL_HOOK_EFER_SYSCALL,
-    SYSCALL_HOOK_EFER_SYSRET,
-
-    CPUID_INSTRUCTION_EXECUTION,
-
-    RDMSR_INSTRUCTION_EXECUTION,
-    WRMSR_INSTRUCTION_EXECUTION,
-
-    IN_INSTRUCTION_EXECUTION,
-    OUT_INSTRUCTION_EXECUTION,
-
-    EXCEPTION_OCCURRED,
-    EXTERNAL_INTERRUPT_OCCURRED,
-
-    DEBUG_REGISTERS_ACCESSED,
-
-    TSC_INSTRUCTION_EXECUTION,
-    PMC_INSTRUCTION_EXECUTION,
-
-    VMCALL_INSTRUCTION_EXECUTION,
-
-} DEBUGGER_EVENT_TYPE_ENUM;
-
-/**
- * @brief Type of Actions
- *
- */
-typedef enum _DEBUGGER_EVENT_ACTION_TYPE_ENUM
-{
-    BREAK_TO_DEBUGGER,
-    RUN_SCRIPT,
-    RUN_CUSTOM_CODE
-
-} DEBUGGER_EVENT_ACTION_TYPE_ENUM;
-
-/**
- * @brief Type of handling !syscall or !sysret
- *
- */
-typedef enum _DEBUGGER_EVENT_SYSCALL_SYSRET_TYPE
-{
-    DEBUGGER_EVENT_SYSCALL_SYSRET_SAFE_ACCESS_MEMORY = 0,
-    DEBUGGER_EVENT_SYSCALL_SYSRET_HANDLE_ALL_UD      = 1,
-
-} DEBUGGER_EVENT_SYSCALL_SYSRET_TYPE;
+//////////////////////////////////////////////////
+//               Event Details                  //
+//////////////////////////////////////////////////
 
 /**
  * @brief Reason for error in parsing commands
@@ -2221,605 +1802,3 @@ typedef struct _DEBUGGEE_EVENT_AND_ACTION_HEADER_FOR_REMOTE_PACKET
 
 } DEBUGGEE_EVENT_AND_ACTION_HEADER_FOR_REMOTE_PACKET,
     *PDEBUGGEE_EVENT_AND_ACTION_HEADER_FOR_REMOTE_PACKET;
-
-//////////////////////////////////////////////////
-//		    	Debugger Success Codes          //
-//////////////////////////////////////////////////
-
-/**
- * @brief General value to indicate that the operation or
- * request was successful
- *
- */
-#define DEBUGGER_OPERATION_WAS_SUCCESSFULL 0xFFFFFFFF
-
-//////////////////////////////////////////////////
-//		    	Debugger Error Codes            //
-//////////////////////////////////////////////////
-
-/**
- * @brief error, the tag not exist
- *
- */
-#define DEBUGGER_ERROR_TAG_NOT_EXISTS 0xc0000000
-
-/**
- * @brief error, invalid type of action
- *
- */
-#define DEBUGGER_ERROR_INVALID_ACTION_TYPE 0xc0000001
-
-/**
- * @brief error, the action buffer size is invalid
- *
- */
-#define DEBUGGER_ERROR_ACTION_BUFFER_SIZE_IS_ZERO 0xc0000002
-
-/**
- * @brief error, the event type is unknown
- *
- */
-#define DEBUGGER_ERROR_EVENT_TYPE_IS_INVALID 0xc0000003
-
-/**
- * @brief error, enable to create event
- *
- */
-#define DEBUGGER_ERROR_UNABLE_TO_CREATE_EVENT 0xc0000004
-
-/**
- * @brief error, invalid address specified for debugger
- *
- */
-#define DEBUGGER_ERROR_INVALID_ADDRESS 0xc0000005
-
-/**
- * @brief error, the core id is invalid
- *
- */
-#define DEBUGGER_ERROR_INVALID_CORE_ID 0xc0000006
-
-/**
- * @brief error, the index is greater than 32 in !exception command
- *
- */
-#define DEBUGGER_ERROR_EXCEPTION_INDEX_EXCEED_FIRST_32_ENTRIES 0xc0000007
-
-/**
- * @brief error, the index for !interrupt command is not between 32 to 256
- *
- */
-#define DEBUGGER_ERROR_INTERRUPT_INDEX_IS_NOT_VALID 0xc0000008
-
-/**
- * @brief error, unable to hide the debugger and enter to transparent-mode
- *
- */
-#define DEBUGGER_ERROR_UNABLE_TO_HIDE_OR_UNHIDE_DEBUGGER 0xc0000009
-
-/**
- * @brief error, the debugger is already in transparent-mode
- *
- */
-#define DEBUGGER_ERROR_DEBUGGER_ALREADY_UHIDE 0xc000000a
-
-/**
- * @brief error, invalid parameters in !e* e* commands
- *
- */
-#define DEBUGGER_ERROR_EDIT_MEMORY_STATUS_INVALID_PARAMETER 0xc000000b
-
-/**
- * @brief error, an invalid address is specified based on current cr3
- * in !e* or e* commands
- *
- */
-#define DEBUGGER_ERROR_EDIT_MEMORY_STATUS_INVALID_ADDRESS_BASED_ON_CURRENT_PROCESS \
-    0xc000000c
-
-/**
- * @brief error, an invalid address is specified based on anotehr process's cr3
- * in !e* or e* commands
- *
- */
-#define DEBUGGER_ERROR_EDIT_MEMORY_STATUS_INVALID_ADDRESS_BASED_ON_OTHER_PROCESS \
-    0xc000000d
-
-/**
- * @brief error, invalid tag for 'events' command (tag id is unknown for kernel)
- *
- */
-#define DEBUGGER_ERROR_MODIFY_EVENTS_INVALID_TAG 0xc000000e
-
-/**
- * @brief error, type of action (enable/disable/clear) is wrong
- *
- */
-#define DEBUGGER_ERROR_MODIFY_EVENTS_INVALID_TYPE_OF_ACTION 0xc000000f
-
-/**
- * @brief error, invalid parameters steppings actions
- *
- */
-#define DEBUGGER_ERROR_STEPPING_INVALID_PARAMETER 0xc0000010
-
-/**
- * @brief error, thread is invalid (not found) or disabled in
- * stepping (step-in & step-out) requests
- *
- */
-#define DEBUGGER_ERROR_STEPPINGS_EITHER_THREAD_NOT_FOUND_OR_DISABLED 0xc0000011
-
-/**
- * @brief error, baud rate is invalid
- *
- */
-#define DEBUGGER_ERROR_PREPARING_DEBUGGEE_INVALID_BAUDRATE 0xc0000012
-
-/**
- * @brief error, serial port address is invalid
- *
- */
-#define DEBUGGER_ERROR_PREPARING_DEBUGGEE_INVALID_SERIAL_PORT 0xc0000013
-
-/**
- * @brief error, invalid core selected in changing core in remote debuggee
- *
- */
-#define DEBUGGER_ERROR_PREPARING_DEBUGGEE_INVALID_CORE_IN_REMOTE_DEBUGGE \
-    0xc0000014
-
-/**
- * @brief error, invalid process selected in changing process in remote debuggee
- *
- */
-#define DEBUGGER_ERROR_PREPARING_DEBUGGEE_UNABLE_TO_SWITCH_TO_NEW_PROCESS \
-    0xc0000015
-
-/**
- * @brief error, unable to run script in remote debuggee
- *
- */
-#define DEBUGGER_ERROR_PREPARING_DEBUGGEE_TO_RUN_SCRIPT 0xc0000016
-
-/**
- * @brief error, invalid register number
- *
- */
-#define DEBUGGER_ERROR_INVALID_REGISTER_NUMBER 0xc0000017
-
-/**
- * @brief error, maximum pools were used without continueing debuggee
- *
- */
-#define DEBUGGER_ERROR_MAXIMUM_BREAKPOINT_WITHOUT_CONTINUE 0xc0000018
-
-/**
- * @brief error, breakpoint already exists on the target address
- *
- */
-#define DEBUGGER_ERROR_BREAKPOINT_ALREADY_EXISTS_ON_THE_ADDRESS 0xc0000019
-
-/**
- * @brief error, breakpoint id not found
- *
- */
-#define DEBUGGER_ERROR_BREAKPOINT_ID_NOT_FOUND 0xc000001a
-
-/**
- * @brief error, breakpoint already disabled
- *
- */
-#define DEBUGGER_ERROR_BREAKPOINT_ALREADY_DISABLED 0xc000001b
-
-/**
- * @brief error, breakpoint already enabled
- *
- */
-#define DEBUGGER_ERROR_BREAKPOINT_ALREADY_ENABLED 0xc000001c
-
-/**
- * @brief error, memory type is invalid
- *
- */
-#define DEBUGGER_ERROR_MEMORY_TYPE_INVALID 0xc000001d
-
-/**
- * @brief error, the process id is invalid
- *
- */
-#define DEBUGGER_ERROR_INVALID_PROCESS_ID 0xc000001e
-
-/**
- * @brief error, for event specific reasons the event is not
- * applied
- *
- */
-#define DEBUGGER_ERROR_EVENT_IS_NOT_APPLIED 0xc000001f
-
-/**
- * @brief error, for process switch or process details, invalid parameter
- *
- */
-#define DEBUGGER_ERROR_DETAILS_OR_SWITCH_PROCESS_INVALID_PARAMETER 0xc0000020
-
-/**
- * @brief error, for thread switch or thread details, invalid parameter
- *
- */
-#define DEBUGGER_ERROR_DETAILS_OR_SWITCH_THREAD_INVALID_PARAMETER 0xc0000021
-
-/**
- * @brief error, maximum breakpoint for a single page is hit
- *
- */
-#define DEBUGGER_ERROR_MAXIMUM_BREAKPOINT_FOR_A_SINGLE_PAGE_IS_HIT 0xc0000022
-
-/**
- * @brief error, there is no pre-allocated buffer
- *
- */
-#define DEBUGGER_ERROR_PRE_ALLOCATED_BUFFER_IS_EMPTY 0xc0000023
-
-/**
- * @brief error, in the EPT handler, it could not split the 2MB pages to
- * 512 entries of 4 KB pages
- *
- */
-#define DEBUGGER_ERROR_EPT_COULD_NOT_SPLIT_THE_LARGE_PAGE_TO_4KB_PAGES 0xc0000024
-
-/**
- * @brief error, failed to get PML1 entry of the target address
- *
- */
-#define DEBUGGER_ERROR_EPT_FAILED_TO_GET_PML1_ENTRY_OF_TARGET_ADDRESS 0xc0000025
-
-/**
- * @brief error, multiple EPT Hooks or Monitors are applied on a single page
- *
- */
-#define DEBUGGER_ERROR_EPT_MULTIPLE_HOOKS_IN_A_SINGLE_PAGE 0xc0000026
-
-/**
- * @brief error, could not build the EPT Hook
- *
- */
-#define DEBUGGER_ERROR_COULD_NOT_BUILD_THE_EPT_HOOK 0xc0000027
-
-/**
- * @brief error, could not find the type of allocation
- *
- */
-#define DEBUGGER_ERROR_COULD_NOT_FIND_ALLOCATION_TYPE 0xc0000028
-
-/**
- * @brief error, could not find the index of test query
- *
- */
-#define DEBUGGER_ERROR_INVALID_TEST_QUERY_INDEX 0xc0000029
-
-/**
- * @brief error, failed to attach to the target user-mode process
- *
- */
-#define DEBUGGER_ERROR_UNABLE_TO_ATTACH_TO_TARGET_USER_MODE_PROCESS 0xc000002a
-
-/**
- * @brief error, failed to remove hooks as entrypoint is not reached yet
- * @details The caller of this functionality should keep sending the previous
- * IOCTL until the hook is remove successfully
- *
- */
-#define DEBUGGER_ERROR_UNABLE_TO_REMOVE_HOOKS_ENTRYPOINT_NOT_REACHED 0xc000002b
-
-/**
- * @brief error, could not remove the previous hook
- *
- */
-#define DEBUGGER_ERROR_UNABLE_TO_REMOVE_HOOKS 0xc000002c
-
-/**
- * @brief error, the needed routines for debugging is not initialized
- *
- */
-#define DEBUGGER_ERROR_FUNCTIONS_FOR_INITIALIZING_PEB_ADDRESSES_ARE_NOT_INITIALIZED 0xc000002d
-
-/**
- * @brief error, unable to get 32-bit or 64-bit of the target process
- *
- */
-#define DEBUGGER_ERROR_UNABLE_TO_DETECT_32_BIT_OR_64_BIT_PROCESS 0xc000002e
-
-/**
- * @brief error, unable to kill the target process
- *
- */
-#define DEBUGGER_ERROR_UNABLE_TO_KILL_THE_PROCESS 0xc000002f
-
-/**
- * @brief error, invalid thread debugging token
- *
- */
-#define DEBUGGER_ERROR_INVALID_THREAD_DEBUGGING_TOKEN 0xc0000030
-
-/**
- * @brief error, unable to pause the process's threads
- *
- */
-#define DEBUGGER_ERROR_UNABLE_TO_PAUSE_THE_PROCESS_THREADS 0xc0000031
-
-/**
- * @brief error, user debugger already attached to this process
- *
- */
-#define DEBUGGER_ERROR_UNABLE_TO_ATTACH_TO_AN_ALREADY_ATTACHED_PROCESS 0xc0000032
-
-/**
- * @brief error, the user debugger is not attached to the target process
- *
- */
-#define DEBUGGER_ERROR_THE_USER_DEBUGGER_NOT_ATTACHED_TO_THE_PROCESS 0xc0000033
-
-/**
- * @brief error, cannot detach from the process as there are paused threads
- *
- */
-#define DEBUGGER_ERROR_UNABLE_TO_DETACH_AS_THERE_ARE_PAUSED_THREADS 0xc0000034
-
-/**
- * @brief error, cannot switch to new thread as the process id or thread id is not found
- *
- */
-#define DEBUGGER_ERROR_UNABLE_TO_SWITCH_PROCESS_ID_OR_THREAD_ID_IS_INVALID 0xc0000035
-
-/**
- * @brief error, cannot switch to new thread the process doesn't contain an active thread
- *
- */
-#define DEBUGGER_ERROR_UNABLE_TO_SWITCH_THERE_IS_NO_THREAD_ON_THE_PROCESS 0xc0000036
-
-/**
- * @brief error, unable to get modules
- *
- */
-#define DEBUGGER_ERROR_UNABLE_TO_GET_MODULES_OF_THE_PROCESS 0xc0000037
-
-/**
- * @brief error, unable to get the callstack
- *
- */
-#define DEBUGGER_ERROR_UNABLE_TO_GET_CALLSTACK 0xc0000038
-
-/**
- * @brief error, unable to query count of processes or threads
- *
- */
-#define DEBUGGER_ERROR_UNABLE_TO_QUERY_COUNT_OF_PROCESSES_OR_THREADS 0xc0000039
-
-//
-// WHEN YOU ADD ANYTHING TO THIS LIST OF ERRORS, THEN
-// MAKE SURE TO ADD AN ERROR MESSAGE TO ShowErrorMessage(UINT32 Error)
-// FUNCTION
-//
-
-//////////////////////////////////////////////////
-//                   IOCTLs                     //
-//////////////////////////////////////////////////
-
-/**
- * @brief ioctl, register a new event
- *
- */
-#define IOCTL_REGISTER_EVENT \
-    CTL_CODE(FILE_DEVICE_UNKNOWN, 0x800, METHOD_BUFFERED, FILE_ANY_ACCESS)
-
-/**
- * @brief ioctl, irp pending mechanism for reading from message tracing buffers
- *
- */
-#define IOCTL_RETURN_IRP_PENDING_PACKETS_AND_DISALLOW_IOCTL \
-    CTL_CODE(FILE_DEVICE_UNKNOWN, 0x801, METHOD_BUFFERED, FILE_ANY_ACCESS)
-
-/**
- * @brief ioctl, to terminate vmx and exit form debugger
- *
- */
-#define IOCTL_TERMINATE_VMX \
-    CTL_CODE(FILE_DEVICE_UNKNOWN, 0x802, METHOD_BUFFERED, FILE_ANY_ACCESS)
-
-/**
- * @brief ioctl, request to read memory
- *
- */
-#define IOCTL_DEBUGGER_READ_MEMORY \
-    CTL_CODE(FILE_DEVICE_UNKNOWN, 0x803, METHOD_BUFFERED, FILE_ANY_ACCESS)
-
-/**
- * @brief ioctl, request to read or write on a speical MSR
- *
- */
-#define IOCTL_DEBUGGER_READ_OR_WRITE_MSR \
-    CTL_CODE(FILE_DEVICE_UNKNOWN, 0x804, METHOD_BUFFERED, FILE_ANY_ACCESS)
-
-/**
- * @brief ioctl, request to read page table entries
- *
- */
-#define IOCTL_DEBUGGER_READ_PAGE_TABLE_ENTRIES_DETAILS \
-    CTL_CODE(FILE_DEVICE_UNKNOWN, 0x805, METHOD_BUFFERED, FILE_ANY_ACCESS)
-
-/**
- * @brief ioctl, register an event
- *
- */
-#define IOCTL_DEBUGGER_REGISTER_EVENT \
-    CTL_CODE(FILE_DEVICE_UNKNOWN, 0x806, METHOD_BUFFERED, FILE_ANY_ACCESS)
-
-/**
- * @brief ioctl, add action to event
- *
- */
-#define IOCTL_DEBUGGER_ADD_ACTION_TO_EVENT \
-    CTL_CODE(FILE_DEVICE_UNKNOWN, 0x807, METHOD_BUFFERED, FILE_ANY_ACCESS)
-
-/**
- * @brief ioctl, request to enable or disable transparent-mode
- *
- */
-#define IOCTL_DEBUGGER_HIDE_AND_UNHIDE_TO_TRANSPARENT_THE_DEBUGGER \
-    CTL_CODE(FILE_DEVICE_UNKNOWN, 0x808, METHOD_BUFFERED, FILE_ANY_ACCESS)
-
-/**
- * @brief ioctl, for !va2pa and !pa2va commands
- *
- */
-#define IOCTL_DEBUGGER_VA2PA_AND_PA2VA_COMMANDS \
-    CTL_CODE(FILE_DEVICE_UNKNOWN, 0x809, METHOD_BUFFERED, FILE_ANY_ACCESS)
-
-/**
- * @brief ioctl, request to edit virtual and physical memory
- *
- */
-#define IOCTL_DEBUGGER_EDIT_MEMORY \
-    CTL_CODE(FILE_DEVICE_UNKNOWN, 0x80a, METHOD_BUFFERED, FILE_ANY_ACCESS)
-
-/**
- * @brief ioctl, request to search virtual and physical memory
- *
- */
-#define IOCTL_DEBUGGER_SEARCH_MEMORY \
-    CTL_CODE(FILE_DEVICE_UNKNOWN, 0x80b, METHOD_BUFFERED, FILE_ANY_ACCESS)
-
-/**
- * @brief ioctl, request to modify an event (enable/disable/clear)
- *
- */
-#define IOCTL_DEBUGGER_MODIFY_EVENTS \
-    CTL_CODE(FILE_DEVICE_UNKNOWN, 0x80c, METHOD_BUFFERED, FILE_ANY_ACCESS)
-
-/**
- * @brief ioctl, flush the kernel buffers
- *
- */
-#define IOCTL_DEBUGGER_FLUSH_LOGGING_BUFFERS \
-    CTL_CODE(FILE_DEVICE_UNKNOWN, 0x80d, METHOD_BUFFERED, FILE_ANY_ACCESS)
-
-/**
- * @brief ioctl, attach or detach user-mode processes
- *
- */
-#define IOCTL_DEBUGGER_ATTACH_DETACH_USER_MODE_PROCESS \
-    CTL_CODE(FILE_DEVICE_UNKNOWN, 0x80e, METHOD_BUFFERED, FILE_ANY_ACCESS)
-
-/**
- * @brief ioctl, print states (Deprecated)
- *
- *
- */
-#define IOCTL_DEBUGGER_PRINT \
-    CTL_CODE(FILE_DEVICE_UNKNOWN, 0x80f, METHOD_BUFFERED, FILE_ANY_ACCESS)
-
-/**
- * @brief ioctl, prepare debuggee
- *
- */
-#define IOCTL_PREPARE_DEBUGGEE \
-    CTL_CODE(FILE_DEVICE_UNKNOWN, 0x810, METHOD_BUFFERED, FILE_ANY_ACCESS)
-
-/**
- * @brief ioctl, pause and halt the system
- *
- */
-#define IOCTL_PAUSE_PACKET_RECEIVED \
-    CTL_CODE(FILE_DEVICE_UNKNOWN, 0x811, METHOD_BUFFERED, FILE_ANY_ACCESS)
-
-/**
- * @brief ioctl, send a signal that execution of command finished
- *
- */
-#define IOCTL_SEND_SIGNAL_EXECUTION_IN_DEBUGGEE_FINISHED \
-    CTL_CODE(FILE_DEVICE_UNKNOWN, 0x812, METHOD_BUFFERED, FILE_ANY_ACCESS)
-
-/**
- * @brief ioctl, send user-mode messages to the debugger
- *
- */
-#define IOCTL_SEND_USERMODE_MESSAGES_TO_DEBUGGER \
-    CTL_CODE(FILE_DEVICE_UNKNOWN, 0x813, METHOD_BUFFERED, FILE_ANY_ACCESS)
-
-/**
- * @brief ioctl, send general buffer from debuggee to debugger
- *
- */
-#define IOCTL_SEND_GENERAL_BUFFER_FROM_DEBUGGEE_TO_DEBUGGER \
-    CTL_CODE(FILE_DEVICE_UNKNOWN, 0x814, METHOD_BUFFERED, FILE_ANY_ACCESS)
-
-/**
- * @brief ioctl, collects a buffer from kernel-side testing informations
- *
- */
-#define IOCTL_SEND_GET_KERNEL_SIDE_TEST_INFORMATION \
-    CTL_CODE(FILE_DEVICE_UNKNOWN, 0x815, METHOD_BUFFERED, FILE_ANY_ACCESS)
-
-/**
- * @brief ioctl, to perform kernel-side tests
- *
- */
-#define IOCTL_PERFROM_KERNEL_SIDE_TESTS \
-    CTL_CODE(FILE_DEVICE_UNKNOWN, 0x816, METHOD_BUFFERED, FILE_ANY_ACCESS)
-
-/**
- * @brief ioctl, to reserve pre-allocated pools
- *
- */
-#define IOCTL_RESERVE_PRE_ALLOCATED_POOLS \
-    CTL_CODE(FILE_DEVICE_UNKNOWN, 0x817, METHOD_BUFFERED, FILE_ANY_ACCESS)
-
-/**
- * @brief ioctl, to send user debugger commands
- *
- */
-#define IOCTL_SEND_USER_DEBUGGER_COMMANDS \
-    CTL_CODE(FILE_DEVICE_UNKNOWN, 0x818, METHOD_BUFFERED, FILE_ANY_ACCESS)
-
-/**
- * @brief ioctl, to get active threads/processes that are debugging
- *
- */
-#define IOCTL_GET_DETAIL_OF_ACTIVE_THREADS_AND_PROCESSES \
-    CTL_CODE(FILE_DEVICE_UNKNOWN, 0x819, METHOD_BUFFERED, FILE_ANY_ACCESS)
-
-/**
- * @brief ioctl, to get user mode modules details
- *
- */
-#define IOCTL_GET_USER_MODE_MODULE_DETAILS \
-    CTL_CODE(FILE_DEVICE_UNKNOWN, 0x81a, METHOD_BUFFERED, FILE_ANY_ACCESS)
-
-/**
- * @brief ioctl, query count of active threads or processes
- *
- */
-#define IOCTL_QUERY_COUNT_OF_ACTIVE_PROCESSES_OR_THREADS \
-    CTL_CODE(FILE_DEVICE_UNKNOWN, 0x81b, METHOD_BUFFERED, FILE_ANY_ACCESS)
-
-/**
- * @brief ioctl, to get list threads/processes
- *
- */
-#define IOCTL_GET_LIST_OF_THREADS_AND_PROCESSES \
-    CTL_CODE(FILE_DEVICE_UNKNOWN, 0x81c, METHOD_BUFFERED, FILE_ANY_ACCESS)
-
-/**
- * @brief ioctl, query the current process details
- *
- */
-#define IOCTL_QUERY_CURRENT_PROCESS \
-    CTL_CODE(FILE_DEVICE_UNKNOWN, 0x81d, METHOD_BUFFERED, FILE_ANY_ACCESS)
-
-/**
- * @brief ioctl, query the current thread details
- *
- */
-#define IOCTL_QUERY_CURRENT_THREAD \
-    CTL_CODE(FILE_DEVICE_UNKNOWN, 0x81e, METHOD_BUFFERED, FILE_ANY_ACCESS)
