+++ hyperdbg/include/SDK/Headers/Ioctls.h
@@ -1,0 +1,234 @@
+/**
+ * @file Ioctls.h
+ * @author Sina Karvandi (sina@hyperdbg.org)
+ * @brief HyperDbg's SDK IOCTL codes
+ * @details This file contains definitions of IOCTLs used in HyperDbg
+ * @version 0.2
+ * @date 2022-06-24
+ *
+ * @copyright This project is released under the GNU Public License v3.
+ *
+ */
+#pragma once
+
+//////////////////////////////////////////////////
+//                   IOCTLs                     //
+//////////////////////////////////////////////////
+
+/**
+ * @brief ioctl, register a new event
+ *
+ */
+#define IOCTL_REGISTER_EVENT \
+    CTL_CODE(FILE_DEVICE_UNKNOWN, 0x800, METHOD_BUFFERED, FILE_ANY_ACCESS)
+
+/**
+ * @brief ioctl, irp pending mechanism for reading from message tracing buffers
+ *
+ */
+#define IOCTL_RETURN_IRP_PENDING_PACKETS_AND_DISALLOW_IOCTL \
+    CTL_CODE(FILE_DEVICE_UNKNOWN, 0x801, METHOD_BUFFERED, FILE_ANY_ACCESS)
+
+/**
+ * @brief ioctl, to terminate vmx and exit form debugger
+ *
+ */
+#define IOCTL_TERMINATE_VMX \
+    CTL_CODE(FILE_DEVICE_UNKNOWN, 0x802, METHOD_BUFFERED, FILE_ANY_ACCESS)
+
+/**
+ * @brief ioctl, request to read memory
+ *
+ */
+#define IOCTL_DEBUGGER_READ_MEMORY \
+    CTL_CODE(FILE_DEVICE_UNKNOWN, 0x803, METHOD_BUFFERED, FILE_ANY_ACCESS)
+
+/**
+ * @brief ioctl, request to read or write on a speical MSR
+ *
+ */
+#define IOCTL_DEBUGGER_READ_OR_WRITE_MSR \
+    CTL_CODE(FILE_DEVICE_UNKNOWN, 0x804, METHOD_BUFFERED, FILE_ANY_ACCESS)
+
+/**
+ * @brief ioctl, request to read page table entries
+ *
+ */
+#define IOCTL_DEBUGGER_READ_PAGE_TABLE_ENTRIES_DETAILS \
+    CTL_CODE(FILE_DEVICE_UNKNOWN, 0x805, METHOD_BUFFERED, FILE_ANY_ACCESS)
+
+/**
+ * @brief ioctl, register an event
+ *
+ */
+#define IOCTL_DEBUGGER_REGISTER_EVENT \
+    CTL_CODE(FILE_DEVICE_UNKNOWN, 0x806, METHOD_BUFFERED, FILE_ANY_ACCESS)
+
+/**
+ * @brief ioctl, add action to event
+ *
+ */
+#define IOCTL_DEBUGGER_ADD_ACTION_TO_EVENT \
+    CTL_CODE(FILE_DEVICE_UNKNOWN, 0x807, METHOD_BUFFERED, FILE_ANY_ACCESS)
+
+/**
+ * @brief ioctl, request to enable or disable transparent-mode
+ *
+ */
+#define IOCTL_DEBUGGER_HIDE_AND_UNHIDE_TO_TRANSPARENT_THE_DEBUGGER \
+    CTL_CODE(FILE_DEVICE_UNKNOWN, 0x808, METHOD_BUFFERED, FILE_ANY_ACCESS)
+
+/**
+ * @brief ioctl, for !va2pa and !pa2va commands
+ *
+ */
+#define IOCTL_DEBUGGER_VA2PA_AND_PA2VA_COMMANDS \
+    CTL_CODE(FILE_DEVICE_UNKNOWN, 0x809, METHOD_BUFFERED, FILE_ANY_ACCESS)
+
+/**
+ * @brief ioctl, request to edit virtual and physical memory
+ *
+ */
+#define IOCTL_DEBUGGER_EDIT_MEMORY \
+    CTL_CODE(FILE_DEVICE_UNKNOWN, 0x80a, METHOD_BUFFERED, FILE_ANY_ACCESS)
+
+/**
+ * @brief ioctl, request to search virtual and physical memory
+ *
+ */
+#define IOCTL_DEBUGGER_SEARCH_MEMORY \
+    CTL_CODE(FILE_DEVICE_UNKNOWN, 0x80b, METHOD_BUFFERED, FILE_ANY_ACCESS)
+
+/**
+ * @brief ioctl, request to modify an event (enable/disable/clear)
+ *
+ */
+#define IOCTL_DEBUGGER_MODIFY_EVENTS \
+    CTL_CODE(FILE_DEVICE_UNKNOWN, 0x80c, METHOD_BUFFERED, FILE_ANY_ACCESS)
+
+/**
+ * @brief ioctl, flush the kernel buffers
+ *
+ */
+#define IOCTL_DEBUGGER_FLUSH_LOGGING_BUFFERS \
+    CTL_CODE(FILE_DEVICE_UNKNOWN, 0x80d, METHOD_BUFFERED, FILE_ANY_ACCESS)
+
+/**
+ * @brief ioctl, attach or detach user-mode processes
+ *
+ */
+#define IOCTL_DEBUGGER_ATTACH_DETACH_USER_MODE_PROCESS \
+    CTL_CODE(FILE_DEVICE_UNKNOWN, 0x80e, METHOD_BUFFERED, FILE_ANY_ACCESS)
+
+/**
+ * @brief ioctl, print states (Deprecated)
+ *
+ *
+ */
+#define IOCTL_DEBUGGER_PRINT \
+    CTL_CODE(FILE_DEVICE_UNKNOWN, 0x80f, METHOD_BUFFERED, FILE_ANY_ACCESS)
+
+/**
+ * @brief ioctl, prepare debuggee
+ *
+ */
+#define IOCTL_PREPARE_DEBUGGEE \
+    CTL_CODE(FILE_DEVICE_UNKNOWN, 0x810, METHOD_BUFFERED, FILE_ANY_ACCESS)
+
+/**
+ * @brief ioctl, pause and halt the system
+ *
+ */
+#define IOCTL_PAUSE_PACKET_RECEIVED \
+    CTL_CODE(FILE_DEVICE_UNKNOWN, 0x811, METHOD_BUFFERED, FILE_ANY_ACCESS)
+
+/**
+ * @brief ioctl, send a signal that execution of command finished
+ *
+ */
+#define IOCTL_SEND_SIGNAL_EXECUTION_IN_DEBUGGEE_FINISHED \
+    CTL_CODE(FILE_DEVICE_UNKNOWN, 0x812, METHOD_BUFFERED, FILE_ANY_ACCESS)
+
+/**
+ * @brief ioctl, send user-mode messages to the debugger
+ *
+ */
+#define IOCTL_SEND_USERMODE_MESSAGES_TO_DEBUGGER \
+    CTL_CODE(FILE_DEVICE_UNKNOWN, 0x813, METHOD_BUFFERED, FILE_ANY_ACCESS)
+
+/**
+ * @brief ioctl, send general buffer from debuggee to debugger
+ *
+ */
+#define IOCTL_SEND_GENERAL_BUFFER_FROM_DEBUGGEE_TO_DEBUGGER \
+    CTL_CODE(FILE_DEVICE_UNKNOWN, 0x814, METHOD_BUFFERED, FILE_ANY_ACCESS)
+
+/**
+ * @brief ioctl, collects a buffer from kernel-side testing informations
+ *
+ */
+#define IOCTL_SEND_GET_KERNEL_SIDE_TEST_INFORMATION \
+    CTL_CODE(FILE_DEVICE_UNKNOWN, 0x815, METHOD_BUFFERED, FILE_ANY_ACCESS)
+
+/**
+ * @brief ioctl, to perform kernel-side tests
+ *
+ */
+#define IOCTL_PERFROM_KERNEL_SIDE_TESTS \
+    CTL_CODE(FILE_DEVICE_UNKNOWN, 0x816, METHOD_BUFFERED, FILE_ANY_ACCESS)
+
+/**
+ * @brief ioctl, to reserve pre-allocated pools
+ *
+ */
+#define IOCTL_RESERVE_PRE_ALLOCATED_POOLS \
+    CTL_CODE(FILE_DEVICE_UNKNOWN, 0x817, METHOD_BUFFERED, FILE_ANY_ACCESS)
+
+/**
+ * @brief ioctl, to send user debugger commands
+ *
+ */
+#define IOCTL_SEND_USER_DEBUGGER_COMMANDS \
+    CTL_CODE(FILE_DEVICE_UNKNOWN, 0x818, METHOD_BUFFERED, FILE_ANY_ACCESS)
+
+/**
+ * @brief ioctl, to get active threads/processes that are debugging
+ *
+ */
+#define IOCTL_GET_DETAIL_OF_ACTIVE_THREADS_AND_PROCESSES \
+    CTL_CODE(FILE_DEVICE_UNKNOWN, 0x819, METHOD_BUFFERED, FILE_ANY_ACCESS)
+
+/**
+ * @brief ioctl, to get user mode modules details
+ *
+ */
+#define IOCTL_GET_USER_MODE_MODULE_DETAILS \
+    CTL_CODE(FILE_DEVICE_UNKNOWN, 0x81a, METHOD_BUFFERED, FILE_ANY_ACCESS)
+
+/**
+ * @brief ioctl, query count of active threads or processes
+ *
+ */
+#define IOCTL_QUERY_COUNT_OF_ACTIVE_PROCESSES_OR_THREADS \
+    CTL_CODE(FILE_DEVICE_UNKNOWN, 0x81b, METHOD_BUFFERED, FILE_ANY_ACCESS)
+
+/**
+ * @brief ioctl, to get list threads/processes
+ *
+ */
+#define IOCTL_GET_LIST_OF_THREADS_AND_PROCESSES \
+    CTL_CODE(FILE_DEVICE_UNKNOWN, 0x81c, METHOD_BUFFERED, FILE_ANY_ACCESS)
+
+/**
+ * @brief ioctl, query the current process details
+ *
+ */
+#define IOCTL_QUERY_CURRENT_PROCESS \
+    CTL_CODE(FILE_DEVICE_UNKNOWN, 0x81d, METHOD_BUFFERED, FILE_ANY_ACCESS)
+
+/**
+ * @brief ioctl, query the current thread details
+ *
+ */
+#define IOCTL_QUERY_CURRENT_THREAD \
+    CTL_CODE(FILE_DEVICE_UNKNOWN, 0x81e, METHOD_BUFFERED, FILE_ANY_ACCESS)
