+++ hyperdbg/include/SDK/HyperDbgSdk.h
@@ -1,0 +1,6 @@
+#pragma once
+
+#include "SDK/Headers/Constants.h"
+#include "SDK/Headers/Datatypes.h"
+#include "SDK/Headers/ErrorCodes.h"
+#include "SDK/Headers/Ioctls.h"
