--- hyperdbg/include/SDK/HyperDbgTypes.h
@@ -1,13 +1,0 @@
-/**
- * @file HyperDbgTypes.h
- * @author Sina Karvandi (sina@hyperdbg.org)
- * @brief HyperDbg's SDK data type definitions
- * @details This file contains definitions of structures, enums, etc.
- * used in HyperDbg
- * @version 0.2
- * @date 2022-06-22
- *
- * @copyright This project is released under the GNU Public License v3.
- *
- */
-#pragma once
